ASoC: clsic-vox: removal of vox_set_idle_and_state()
The helper function was no-longer adding value and left the driver in an
inconsistent state if an error occurred.

+++ sound/soc/codecs/clsic-vox.h
@@ -147,7 +147,7 @@ struct clsic_vox {
 	/* drv_state refers to ongoing vox biometric operations only. */
 	unsigned int drv_state;
 	/* Which mode CLSIC is in. */
-	int clsic_mode;
+	int8_t clsic_mode;
 
 	/*
 	 * Mutex to guard the testing and changing of clsic_vox structure
